GPL-3 -> Apache-2
I've not been an active maintainer for many years and there's no reason
to hold on to the GPL-3, given the general advancement of BSD- and
Apache-style licenses over the past decade.

This change only applies to the code I personally wrote. Other authors
retain the right to decide the license to their contributions, which
should be discoverable using `git blame`.